requirements.txt
# Core
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.8
langchain-text-splitters>=0.2.0

# Vector Store
faiss-cpu>=1.7.4

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Environment
python-dotenv>=1.0.0

# Backend API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Utilities
tiktoken>=0.5.0
